Polishing up for Phase 1 second version

+++ Webfuel.App/ClientApp/src/app/features/project/project/project-submission/project-submission.component.ts
@@ -7,32 +7,23 @@ import { StaticDataCache } from 'api/static-data.cache';
 import { FormService } from 'core/form.service';
 import { ConfirmDeleteDialog } from '../../../../shared/dialogs/confirm-delete/confirm-delete.dialog';
 import { UpdateProjectSubmissionDialog } from '../project-submission/update-project-submission/update-project-submission.dialog';
+import { ProjectComponentBase } from '../shared/project-component-base';
+import { CreateProjectSubmissionDialog } from './create-project-submission/create-project-submission.dialog';
 
 @Component({
   selector: 'project-submission',
   templateUrl: './project-submission.component.html'
 })
-export class ProjectSubmissionComponent implements OnInit {
+export class ProjectSubmissionComponent extends ProjectComponentBase {
 
   constructor(
-    private route: ActivatedRoute,
-    private router: Router,
     private formService: FormService,
+    private createProjectSubmissionDialog: CreateProjectSubmissionDialog,
     private updateProjectSubmissionDialog: UpdateProjectSubmissionDialog,
     private confirmDeleteDialog: ConfirmDeleteDialog,
     public projectSubmissionApi: ProjectSubmissionApi,
-    public staticDataCache: StaticDataCache
   ) {
-  }
-
-  ngOnInit() {
-    this.reset(this.route.snapshot.data.project);
-  }
-
-  item!: Project;
-
-  reset(item: Project) {
-    this.item = item;
+    super();
   }
 
   form = new FormGroup({
@@ -47,11 +38,19 @@ export class ProjectSubmissionComponent implements OnInit {
     query.projectId = this.item.id;
   }
 
+  add() {
+    if (this.locked) return;
+    this.createProjectSubmissionDialog.open({ projectId: this.item.id });
+  }
+
+
   edit(item: ProjectSubmission) {
+    if (this.locked) return;
     this.updateProjectSubmissionDialog.open({ projectSubmission: item });
   }
 
   delete(item: ProjectSubmission) {
+    if (this.locked) return;
     this.confirmDeleteDialog.open({ title: "Project Submission" }).subscribe((result) => {
       this.projectSubmissionApi.delete({ id: item.id }, { successGrowl: "Project Submission Deleted" }).subscribe();
     });
